chart/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
asyncpg==0.30.0
dnslib==0.9.24
psycopg2-binary==2.9.10
